refactor: move reusable runner result to its file
